requirements.txt
Django==2.1.2
django-extensions==2.1.3
djangorestframework==3.9.0
psycopg2-binary==2.7.5
factory-boy==2.9.2